Update lexer: adding string literals lexer

+++ crates/guiql/src/lang/ast.rs
@@ -1,7 +1,7 @@
 #[derive(PartialEq, Clone, Debug)]
 pub struct TokenLoc {
-    pub starts_at: usize,
-    pub len: usize,
+    pub starts_at: u32,
+    pub len: u32,
 }
 
 #[derive(PartialEq, Clone, Debug)]
